+++ Runtime/Scripts/Export/Constants.cs
@@ -7,7 +7,7 @@ namespace GLTFast.Export
 {
     static class Constants
     {
-        public const string version = "6.0.0-pre.1";
+        public const string version = "6.0.0-pre.2";
 
         internal const string mimeTypePNG = "image/png";
         internal const string mimeTypeJPG = "image/jpeg";