+++ CHANGELOG.md
@@ -20,6 +20,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `MeshAdded`
   - `EndSceneCompleted`
 
+## [Unreleased]
+### Added
+- (Export) Support for Draco mesh compressed exports
+### Changed
+- Added proper root namespace to all assembly definitions
+
 ## [5.0.4] - 2023-03-30
 ### Fixed
 - Texture transform offset is calculated correctly now