+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [6.0.0-pre.1] - 2023-05-12
+## [6.0.0-pre.2] - 2023-05-25
 ### Added
 - Custom Extension API (`GLTFast.Extensions` namespace)
 - Support for alternative JSON parsing via Newtonsoft JSON